DROTH-1122 Service: Extend PointAssetOperations for new point width limit - Moved some classes and objects from DAO to Service; - Created WidthLimitService.
